[ADD] 프로젝트 파일 업로드

+++ src/main/java/com/codestates/soloproject/v1/repository/CompanyLocationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CompanyLocationRepository extends JpaRepository<CompanyType, Integer> {
+public interface CompanyLocationRepository extends JpaRepository<CompanyType, Long> {
 }
